requirements.txt
# Core dependencies
pyspark>=3.4.0
delta-spark>=2.4.0

# Streamlit App
streamlit>=1.29.0
databricks-sql-connector>=3.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
chispa>=0.9.0  # PySpark testing utilities

# Linting and formatting
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Databricks SDK
databricks-sdk>=0.12.0
databricks-cli>=0.18.0
